Fixes for PERF-406

+++ lib/db/index.ts
@@ -60,7 +60,7 @@ export function initDb() {
       user_id INTEGER NOT NULL REFERENCES users(id),
       account_number TEXT UNIQUE NOT NULL,
       account_type TEXT NOT NULL,
-      balance REAL DEFAULT 0 NOT NULL,
+      balance INTEGER DEFAULT 0 NOT NULL,
       status TEXT DEFAULT 'pending',
       created_at TEXT DEFAULT CURRENT_TIMESTAMP
     );
@@ -69,7 +69,7 @@ export function initDb() {
       id INTEGER PRIMARY KEY AUTOINCREMENT,
       account_id INTEGER NOT NULL REFERENCES accounts(id),
       type TEXT NOT NULL,
-      amount REAL NOT NULL,
+      amount INTEGER NOT NULL,
       description TEXT,
       status TEXT DEFAULT 'pending' NOT NULL,
       created_at TEXT DEFAULT CURRENT_TIMESTAMP,
